Various minor fixes to allow building with clang on OS X.
Fixes issue 3992. 

wxWidgets patch at http://trac.wxwidgets.org/ticket/12332 .


git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@6929 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Common/Src/Misc.cpp
@@ -41,8 +41,9 @@ const char *GetLastErrorMsg()
 // strlen with cropping after size n
 size_t strnlen(const char *s, size_t n)
 {
-  const char *p = (const char *)memchr(s, 0, n);
-  return(p ? p-s : n);
+	const char *p = (const char *)memchr(s, 0, n);
+
+	return p ? (size_t)(p - s) : n;
 }
 #endif
 #if defined(_WIN32) || !(__GNUC__ == 4 && __GNUC_MINOR__ >= 5)
